backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.0
plexapi==4.15.10
httpx==0.27.0
bcrypt==4.1.3
PyJWT==2.8.0
slowapi==0.1.9
python-multipart==0.0.9
ollama==0.3.3
